requirements.txt
flask
torch
torchvision
pillow
gunicorn
gdown